requirements.txt
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
google-adk>=0.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0
quart
